clean_kernel.sh: Update script

+++ clean_kernel.sh
@@ -4,16 +4,20 @@
 export ARCH=arm
 export CROSS_COMPILE=/opt/toolchains/arm-eabi-4.7/bin/arm-eabi-
 
-echo "Start cleaning kernel..."
+if [ -d output ]; then
+    echo "Start cleaning kernel..."
 
-cd output
-make clean && make mrproper
-cd ../
+    cd output
+    make clean && make mrproper
+    cd ../
 
-echo "Remove output directory..."
-rm -rf output
+    echo "Remove output directory..."
+    rm -rf output
 
-echo "Remove output files..."
-rm -rf utilities/AnyKernel2/Kernel.zip Kernel.zip utilities/AnyKernel2/zImage utilities/AnyKernel2/modules/system/lib/modules/pronto/pronto_wlan.ko arch/arm/boot/zImage
+    echo "Remove output files..."
+    rm -rf utilities/AnyKernel2/Kernel.zip Kernel.zip utilities/AnyKernel2/zImage utilities/AnyKernel2/modules/system/lib/modules/pronto/pronto_wlan.ko arch/arm/boot/zImage
 
-echo "Finish!"
+    echo "Finish!"
+else
+   echo "Don't need cleaning."
+fi
